+++ cmp/util.h
@@ -9,7 +9,7 @@
 
 char *_ctimes(char, size_t);
 
-unsigned int countcs(unsigned int, char *, char);
+size_t countcs(unsigned int, char *, char);
 
 void _loop_id(char * const, unsigned int);
